requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-scheduler==0.10.1
environs==8.0.0
frozenlist==1.4.1
icalendar==5.0.11
idna==3.6
magic-filter==1.0.12
marshmallow==3.21.0
multidict==6.0.5
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
yarl==1.9.4
